Login Screen, SignUp Screen, Sign Out functionality
replaced current navigation with navigation Stack, some dependencies may need to be installed on your local machines, but it makes switching screens smoother. as title says login, sign up, and sign out are all functional, and the name in the side menu now pertains to the user logged in as opposed to hard coded jessica jones.